# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "App AI Gateway API",
    "version": "0.1.0",
    "description": "A multi-tenant, provider-native LLM proxy for mobile applications and trusted server backends."
  },
  "servers": [
    {
      "url": "https://gateway.example.com",
      "description": "Replace with your deployed gateway origin"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ManagementBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "agw_mgmt_…",
        "description": "An organization-scoped management API key created by an owner or admin."
      },
      "OperatorSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agw_operator_auth.session_token",
        "description": "The Better Auth operator session cookie. Admin requests also send x-console-request: 1."
      },
      "GatewayBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "A gateway access token, or an application API key for an issuer-less API-key app."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "data": {
                "type": "object",
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "description": "Machine-readable facts about this rejection, present only where the code alone is not actionable. A billing_request_quota_exceeded rejection carries limit, used, and resetAt; an app_rate_limited or app_budget_exhausted rejection carries scope, either user or app."
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/apps/{app}/me": {
      "get": {
        "tags": [
          "Application"
        ],
        "operationId": "getCurrentUser",
        "summary": "Get the current user's limits, spend and block state",
        "security": [
          {
            "GatewayBearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Current user state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "string"
                    },
                    "limits": {
                      "type": "object",
                      "properties": {
                        "requests_today": {
                          "type": [
                            "integer",
                            "null"
                          ],
                          "description": "Requests this user has made so far in the current UTC day. Null when the application sets no per-user limit, in which case requests are not counted at all."
                        },
                        "requests_remaining": {
                          "type": [
                            "integer",
                            "null"
                          ],
                          "description": "What is left of requests_per_day. Null when no daily limit is set."
                        },
                        "requests_per_minute": {
                          "type": [
                            "number",
                            "null"
                          ],
                          "description": "The per-user per-minute limit this app sets. Null means unlimited."
                        },
                        "requests_per_day": {
                          "type": [
                            "number",
                            "null"
                          ],
                          "description": "The per-user per-day limit this app sets. Null means unlimited."
                        },
                        "monthly_cost_usd": {
                          "type": "number",
                          "description": "What this user's traffic has cost so far in the current UTC calendar month."
                        },
                        "monthly_budget_usd": {
                          "type": [
                            "number",
                            "null"
                          ],
                          "description": "The per-user monthly spending budget this app sets. Null means unlimited."
                        },
                        "blocked": {
                          "type": "boolean",
                          "description": "Whether an operator has blocked this user."
                        }
                      },
                      "required": [
                        "requests_today",
                        "requests_remaining",
                        "requests_per_minute",
                        "requests_per_day",
                        "monthly_cost_usd",
                        "monthly_budget_usd",
                        "blocked"
                      ],
                      "description": "The limits the application's operator set on this user, and where the user stands against them. The organization's own plan allowance is not reported here: it is organization-wide, and is reported on the rejection that spends it."
                    }
                  },
                  "required": [
                    "user_id",
                    "limits"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "No billing plan resolves for the organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/apps/{app}/endpoints/{slug}": {
      "post": {
        "tags": [
          "Named endpoints"
        ],
        "operationId": "callNamedEndpoint",
        "summary": "Call a server-configured named endpoint",
        "description": "The endpoint's provider, model, fixed parameters, output cap, and fallback chain come from the application configuration, so an operator can change models without shipping a client release. Responses-style endpoints accept an OpenAI Responses body; transcription-style endpoints accept an OpenAI audio transcription multipart body and may omit the model field. The successful response keeps the serving provider's native format and streaming behaviour.",
        "security": [
          {
            "GatewayBearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "my-app"
            },
            "required": true,
            "name": "app",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]{1,64}$",
              "example": "chat"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Required for gateway-token clients; optional for issuer-less API-key clients."
            },
            "required": false,
            "description": "Required for gateway-token clients; optional for issuer-less API-key clients.",
            "name": "x-app-version",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Optional configured end-user identity for issuer-less API-key applications."
            },
            "required": false,
            "description": "Optional configured end-user identity for issuer-less API-key applications.",
            "name": "x-end-user-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {},
                "description": "OpenAI Responses API body for endpoints whose api_style is responses. The gateway overwrites model and deep-merges the configured params."
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "model": {
                    "type": "string",
                    "description": "Ignored; the gateway sets the configured model."
                  },
                  "prompt": {
                    "type": "string"
                  },
                  "language": {
                    "type": "string"
                  },
                  "response_format": {
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "description": "Body for endpoints whose api_style is transcription."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider-native response. Streaming responses remain streamed.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "No billing plan resolves for the organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated identity is not allowed to perform this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Either the app's own limits or the organization's plan allowance refused the request. The app's limits are checked first and answer app_rate_limited or app_budget_exhausted, carrying scope; the plan allowance answers billing_request_quota_exceeded with limit, used, and the UTC resetAt. Every one of them carries a Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Every configured target failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Every configured target failed, and the last one sent no response headers within the gateway's time-to-first-byte budget.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}